requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<2.0
gunicorn>=23.0,<24.0
psycopg2-binary>=2.9,<3.0
qrcode[pil]>=7.4,<9.0
requests>=2.31,<3.0
python-docx>=1.1,<2.0
Pillow>=10.0,<12.0
python-barcode>=0.15,<1.0
pyzbar>=0.1.9,<1.0
